maybe gracefully handle 401 error
